Fix KeyChain crash
This happens due to a race condition in background threads that use an
activity context to call long-running RPCs in keychain and keystore.

If the activity goes away while services are still bound, its context
will be disposed of and any attempts to unbind those services will raise
an IllegalStateException.

The right thing to do is make these calls using an application context
instead.