Remove hidden attribute in restrictions (closes #1200)
According to the discussion in #1194, Google GSuite  is not following their
own spec that hidden fields should not be shown and set to their
implicit default value. Unhide the field, so user can set it to
the default value.

From the docs (https://developer.android.com/reference/android/content/RestrictionsManager):

     *hidden* type must have a defaultValue and will not be shown to the administrator. It can be used to pass along data that cannot be modified, such as a version code.