Fix binding to external authenticator app on Android 14+
External authenticator apps may need to start background
activities (for example, to allow users enter a pin code), but
starting from Android 14, if the app bound to the service is
targeting Android 14 or higher, it no longer allows the app that
has the service to start a background activity by default. We
need to add flag BIND_ALLOW_ACTIVITY_STARTS to allow the
external authenticator app to start background activities.