Allow disabling VPN confirmation
This is not recommended but people keep annoying for their broken Samsung
automation nonsense that abuses Android APIs for their own ****** instead
of providing proper APIs or trying to improve overall ecosystem. So app
programmers are bugged to make questionable implementation choices like
this to cater to that mess that Samsung created.

Github: closes schwabe/ics-openvpn#1808